requirements.txt
docling
langchain
langchain_groq
python-dotenv
transformers
pinecone
sentence_transformers